+++ e2e-tests/contentful/cypress/integration/gatsby-image.js
@@ -31,8 +31,9 @@ function testGatsbyImage(type, testPlaceholder) {
       })
   }
 
-  cy.get(`[data-cy="${type}"]`).scrollIntoView()
-  // Wait for load
+  cy.get(`[data-cy="${type}"]`).scrollIntoView({ duration: 500 })
+
+  // Wait images for load
   cy.wait(1000)
 
   cy.get(`[data-cy="${type}"]`)
@@ -53,4 +54,19 @@ describe(`gatsby-image`, () => {
   it(`webp`, testConfig, () => testGatsbyImage(`webp`, hasJPEGPlaceholder))
   it(`traced`, testConfig, () => testGatsbyImage(`traced`, hasSVGPlaceholder))
   it(`sqip`, testConfig, () => testGatsbyImage(`sqip`, hasSVGPlaceholder))
+
+  it(`english`, testConfig, () => {
+    testGatsbyImage(`english`, hasJPEGPlaceholder)
+    cy.get(`[data-cy="english"] p strong`).should(
+      "have.text",
+      "Locale - American English (png)"
+    )
+  })
+  it(`german`, testConfig, () => {
+    testGatsbyImage(`german`, hasJPEGPlaceholder)
+    cy.get(`[data-cy="german"] p strong`).should(
+      "have.text",
+      "Locale - German (png)"
+    )
+  })
 })