+++ packages/gatsby-remark-custom-blocks/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.8.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-custom-blocks@3.7.0-next.2...gatsby-remark-custom-blocks@3.8.0-next.0) (2021-06-03)
+
+**Note:** Version bump only for package gatsby-remark-custom-blocks
+
+# [3.7.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-custom-blocks@3.7.0-next.1...gatsby-remark-custom-blocks@3.7.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-remark-custom-blocks
+
+# [3.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-custom-blocks@3.7.0-next.0...gatsby-remark-custom-blocks@3.7.0-next.1) (2021-05-31)
+
+**Note:** Version bump only for package gatsby-remark-custom-blocks
+
 # [3.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-custom-blocks@3.6.0-next.0...gatsby-remark-custom-blocks@3.7.0-next.0) (2021-05-21)
 
 **Note:** Version bump only for package gatsby-remark-custom-blocks